extract support code from spec/integration
