+++ spec/integration/support/index.js
@@ -1,0 +1,8 @@
+export function debugResponse(response) {
+  const request = response.request()
+  return `Status: ${response.status()}, Headers: ${JSON.stringify(request.headers())}, ${request.method().toUpperCase()} ${request.url()} => ${response.rawData()}`
+}
+
+export function errorMessage(response) {
+  return response.rawData ? debugResponse(response) : response
+}
