+++ spec/integration/support/manifest.js
